requirements.txt
streamlit
pandas
numpy
scipy
plotly
google-generativeai
python-dotenv